backend/requirements.txt
Django==5.2.11
djangorestframework==3.15.*
django-cors-headers==4.6.*
psycopg[binary]==3.2.*
gunicorn==23.0.*
whitenoise==6.8.*
python-decouple==3.8
openai>=1.0.0
